move old helpers into session feature
