+++ src/masonite/helpers/__init__.py
@@ -3,4 +3,3 @@ from .optional import Optional as optional
 from .mix import MixHelper
 from .urls import UrlsHelper
 from .compact import Compact as compact
-from .old import old
