+++ src/masonite/sessions/helpers.py
@@ -2,4 +2,5 @@ from ..facades import Session
 
 
 def old(key):
+    """"""
     return Session.get(key) or ""
